Port i18nFormatting test to Jest
