--- __mocks__/i18next.js
@@ -1,10 +1,0 @@
-export default {
-  init: jest.fn(),
-
-  t(key) {
-    if (key === 'utility.or') {
-      return ' or ';
-    }
-    return '';
-  },
-};
